backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.7
python-multipart>=0.0.6
openpyxl>=3.1.2
qrcode[pil]>=7.4.2
pillow>=10.1.0
python-dotenv>=1.0.0
websockets>=12.0
opencv-python-headless>=4.8.0
pyzbar>=0.1.9